requirements.txt
flask
Flask-PyMongo
python-decouple
dnspython
python-dotenv